support HTTP method field in node 0.12
In my last pull request I had missed the method field. I had to make a
hack for backward compatibility, so that HTTPParser knows wether to pass
the string or the numerical index. I also diffed the logs of a small
test app with and without this implementation, and found some other
differences.