Complete the parser for a HTTP request when no request body.
if we are done with headers and there is no implied content, go ahead
and flag the parser as complete.

Conformant HTTP clients will no- include a body with a HTTP request
and fail to include a Transfer-Encoding or Content-Length header.

Reexecuting the parser on some listening sockets will otherwise cause
http-parser-js to hang waiting for more data in node 0.12.x.  I have
reporoduced this behavior during my testing of
[hipache](https://github.com/hipache/hipache) with http-parser-js
monkeypatched in.

This should be able to maintain the other http-parser-js functionality
as far as I can tell.